move income_uniform_sample and MAXIMUM_INCOME_FACTOR to a utils module
